refactor utils #9
